refactor, use ibatis instead of mybatis
