+++ jakarta-ee/student-web-app/.gitignore
@@ -3,3 +3,4 @@ build/
 .idea/
 *.iml
 .vscode/
+/bin/
