+++ jakarta-ee/student-web-app/liberty_config/server.env
@@ -9,4 +9,8 @@ TRUSTED_KEYSTORE_PASSWORD=admin
 
 LTPA_KEY_FILE_SFA=D:\open-liberty\openliberty-25.0.0.4\wlp\usr\servers\jimmyserver\conf\ltpa_sfa.key
 LTPS_KEY_PASSWORD_SFA=admin
+
+
 LTPA_SFA_EXPIRATION=3600
+
+#POD_NAME=local-server-YOURNAME need to change "YOURNAME" in local
