+++ jakarta-ee/student-web-app/resources/ca/on/gov/edu/msfaa/shared/persistence/xml/Test_SqlMap.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE sqlMap
+        PUBLIC "-//ibatis.apache.org//DTD SQL Map 2.0//EN"
+        "http://ibatis.apache.org/dtd/sql-map-2.dtd">
+<sqlMap namespace="com.azure.sample.TestMapper">
+
+<!-- SELECT 操作 -->
+<select id="listStudent" resultClass="ca.on.gov.edu.coreft.StudentProfile">
+    SELECT id, name, email FROM student_profiles
+</select>
+
+<!-- INSERT 操作 -->
+<insert id="addStudent" parameterClass="java.util.Map">
+    INSERT INTO student_profiles (name, email, major) VALUES (#name#, #email#, #major#)
+</insert>
+
+</sqlMap>
