+++ jakarta-ee/student-web-app/resources/log4j.properties
@@ -6,9 +6,9 @@ log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} [%t] %-5p %c !!! %m%n
 
 # File appender configuration
-log4j.appender.file=org.apache.log4j.RollingFileAppender
-log4j.appender.file.File=D:\\temp\\application.log
-log4j.appender.file.MaxFileSize=10MB
-log4j.appender.file.MaxBackupIndex=5
-log4j.appender.file.layout=org.apache.log4j.PatternLayout
-log4j.appender.file.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} [%t] %-5p %c --- %m%n
+log4j.appender.rolling=org.apache.log4j.RollingFileAppender
+log4j.appender.rolling.File=/logs/applog/osap/coreft1617/coreft1617sfa.log
+log4j.appender.rolling.MaxFileSize=100000KB
+log4j.appender.rolling.MaxBackupIndex=10
+log4j.appender.rolling.layout=org.apache.log4j.PatternLayout
+log4j.appender.rolling.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} [%t] %-5p %c --- %m%n
