+++ jakarta-ee/student-web-app/resources/sql-map-config.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE configuration
-        PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-config.dtd">
-<configuration>
-    <environments default="liberty">
-        <environment id="liberty">
-            <transactionManager type="JDBC"/>
-            <dataSource type="JNDI">
-                <property name="data_source" value="jdbc/StudentDB"/>
-            </dataSource>
-        </environment>
-    </environments>
-    <mappers>
-         <mapper resource="persistence/xml/StudentMapper.xml"/>
-    </mappers>
-</configuration>
+<!DOCTYPE sqlMapConfig
+        PUBLIC "-//ibatis.apache.org//DTD SQL Map Config 2.0//EN"
+        "http://ibatis.apache.org/dtd/sql-map-config-2.dtd">
+<sqlMapConfig>
+    <settings useStatementNamespaces="true"/>
+
+    <transactionManager type="JDBC">
+        <dataSource type="JNDI">
+            <property name="DataSource" value="jdbc/StudentDB"/>
+        </dataSource>
+    </transactionManager>
+
+    <sqlMap resource="ca/on/gov/edu/msfaa/shared/persistence/xml/Student_SqlMap.xml"/>
+    <sqlMap resource="ca/on/gov/edu/msfaa/shared/persistence/xml/Test_SqlMap.xml"/>
+    <sqlMap resource="ca/on/gov/edu/msfaa/shared/persistence/xml/MSFAA_MASTER_SqlMap.xml"/>
+</sqlMapConfig>
