+++ jakarta-ee/student-web-app/src/ca/on/gov/edu/coreft/HelloWorldServlet.java
@@ -1,4 +1,4 @@
-package com.azure.sample;
+package ca.on.gov.edu.coreft;
 import org.apache.log4j.Logger;
 
 import java.io.*;
