+++ jakarta-ee/student-web-app/src/ca/on/gov/edu/coreft/StudentProfile.java
@@ -1,4 +1,4 @@
-package com.azure.sample;
+package ca.on.gov.edu.coreft;
 import java.io.Serializable;
 
 public class StudentProfile implements Serializable {
