+++ jakarta-ee/student-web-app/src/ca/on/gov/edu/coreft/filter/CommonHttpServletFilter.java
@@ -1,4 +1,4 @@
-package com.azure.sample.listener;
+package ca.on.gov.edu.coreft.filter;
 
 import javax.servlet.*;
 import javax.servlet.http.HttpServletRequest;
